refactor:refactor cart styles
