+++ src/components/Cart/CartItem/CartItemStyle.scss
@@ -1,9 +1,9 @@
+@import '../../../styles/abstract/mixins';
+
 .cart {
   width: 100%;
   height: 12rem;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
+  @include flexCenter(space-around);
 
   &__name,
   &__price {
@@ -11,13 +11,7 @@
     font-size: 1.4rem;
   }
   &__price {
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-  }
-  &__img {
-    width: 10rem;
-    height: auto;
+    @include flexCenter(start, column);
   }
 
   &__total-price {
@@ -25,10 +19,8 @@
     font-size: 1.6rem;
   }
   &__active {
-    width: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
+    width: 30%;
+    @include flexCenter(space-around);
   }
 }
 
@@ -76,18 +68,8 @@
 }
 @media (max-width: 640px) {
   .cart {
-    &__image img {
-      width: 50%;
-    }
     &__quantity {
       margin: 2rem 0;
-    }
-    &__button,
-    &__name {
-      margin-right: 0;
-    }
-    &__name {
-      text-align: center;
     }
   }
 }
